+++ cpp/include/Matrix_math.h
@@ -1,0 +1,10 @@
+#ifndef MATRIX_MATH_H_INCLUDED
+#define MATRIX_MATH_H_INCLUDED
+
+#include "math/doolittle.h"
+#include "math/ldu.h"
+#include "math/crout.h"
+#include "math/eye.h"
+#include "math/inv.h"
+
+#endif // MATRIX_MATH_H_INCLUDED