+++ cpp/src/Complex_io.cpp
@@ -1,25 +1,30 @@
 #include <cmath>
 #include <cstdio>
 #include <iomanip>
-#include "ComplexNumber.h"
-#include "ComplexNumber_io.h"
+#include "Complex.h"
 using namespace std;
 
 ZMAT_BEGIN_NAMESPACE
 
-istream& operator >>(istream &in, ComplexNumber &num) {
+istream& operator >>(istream &in, Complex &num)
+{
     double real, image;
     in >> real >> image;
     num.setValue(real, image);
     return in;
 }
 
-ostream& operator <<(ostream &out, const ComplexNumber &num) {
+ostream& operator <<(ostream &out, const Complex &num)
+{
     out << num.real();
-    if (fabs(num.image()) > ZMAT_EPS) {
-        if (num.image() <= 0.0) {
+    if (fabs(num.image()) > ZMAT_EPS)
+    {
+        if (num.image() <= 0.0)
+        {
             out << " - ";
-        } else {
+        }
+        else
+        {
             out << " + ";
         }
         out << num.image() << " i";