+++ cpp/test/Test_Decompose.h
@@ -1,0 +1,6 @@
+#ifndef TEST_DECOMPOSE_H_INCLUDED
+#define TEST_DECOMPOSE_H_INCLUDED
+
+void Test_Decompose();
+
+#endif // TEST_DECOMPOSE_H_INCLUDED