+++ cpp/test/Test_Generate.h
@@ -1,0 +1,6 @@
+#ifndef TEST_GENERATE_H
+#define TEST_GENERATE_H
+
+void Test_Generate();
+
+#endif // TEST_GENERATE_H